update to v1.2
- Removed Slashcommand Setup:
The setup via slash commands has been removed. Now, accessing the Interface/Addons tab is required for addon configuration.

- Interface/Addons Tab Setup:
Added an Interface/Addons tab for convenient setup of the addon.

- Code Refactor:
The code has undergone a refactoring process for improved structure and readability.

- Character-Specific Configurations:
Configurations and saved variables are now specific to each character.

- Optimized Event Handling:
Improved performance by dynamically registering and unregistering events when entering or leaving Dalaran.

- Dynamic Engineering Grandmaster Check:
If a player gains Engineering Grandmaster the addon will notice it and start it self, relogging is no longer necessary.

- Custom Raid Mark Selection:
Added the option to select a specific raid mark to be applied to the Auctionhouse Bot.

- Localization Rework:
Localization of the addon has been reworked to align with the recent changes.